requirements.txt
gunicorn==19.0.0
beautifulsoup4==4.6.0
bs4==0.0.1
Flask==0.12.2
Flask-Cors==3.0.3
Jinja2==2.9.6
networkx==1.11
requests==2.13.0
Werkzeug==0.12.2
PyGithub==1.43.3 
